pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minigpt-jax-stories"
version = "0.1.0"
description = "MiniGPT in JAX for short story generation"
authors = [
  { name = "Dr. Sajil C K" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "jax",
  "jaxlib",
  "flax",
  "optax",
  "orbax-checkpoint",
  "numpy",
  "pandas",
  "tiktoken",
  "pyyaml",
  "matplotlib"
]

[tool.setuptools.packages.find]
where = ["src"]